refactor(chat): move db schemas to api package
