+++ packages/api/src/chat/schema.ts
